pyhurdy.py: Add X11 support
This commit adds support for running within x11. Prior to
this, you could only run within one of the supported
framebuffers. What this means is that you no longer
need to be root to run this.

I've also removed some of the debug print statements,
and cleaned up the horrible key event code which is going
away soon.

Signed-off-by: Beth Flanagan <elizabeth.flanagan@intel.com>